+++ flask_unchained/bundles/controller/hooks/__init__.py
@@ -4,7 +4,7 @@ from typing import *
 from ..route import Route
 from .register_blueprints_hook import RegisterBlueprintsHook
 from .register_routes_hook import RegisterRoutesHook
-from .register_bundle_template_folders import RegisterBundleTemplateFoldersHook
+from .register_bundle_blueprints_hook import RegisterBundleBlueprintsHook
 
 EndpointName = str
 BundleName = str