+++ flask_unchained/bundles/controller/hooks/register_blueprints_hook.py
@@ -2,8 +2,6 @@ from flask import Blueprint, Flask
 from flask_unchained import AppFactoryHook, Bundle
 from typing import *
 
-from ..utils import get_babel_bundle
-
 
 class RegisterBlueprintsHook(AppFactoryHook):
     bundle_module_name = 'views'
@@ -18,7 +16,6 @@ class RegisterBlueprintsHook(AppFactoryHook):
     _limit_discovery_to_local_declarations = False
 
     def run_hook(self, app: Flask, bundles: List[Type[Bundle]]):
-        self.babel_bundle = get_babel_bundle(bundles)
         super().run_hook(app, bundles)
 
     def process_objects(self, app: Flask, blueprints: List[Blueprint]):
@@ -29,9 +26,6 @@ class RegisterBlueprintsHook(AppFactoryHook):
             url_prefix = (blueprint.url_prefix or '').rstrip('/')
             app.register_blueprint(blueprint, url_prefix=url_prefix)
             self.log_action(blueprint)
-
-            if self.babel_bundle:
-                self.babel_bundle.register_blueprint(app, blueprint)
 
     def collect_from_bundles(self, bundles: List[Type[Bundle]],
                              ) -> List[Blueprint]: