+++ flask_unchained/bundles/controller/hooks/register_bundle_blueprints_hook.py
@@ -4,9 +4,6 @@ from flask.helpers import _endpoint_from_view_func
 from flask_unchained import AppFactoryHook, Bundle
 from typing import List
 
-from ..utils import get_babel_bundle
-
-
 # FIXME test template resolution order when this is used in combination with
 # RegisterBlueprintsHook
 
@@ -78,7 +75,7 @@ class BundleBlueprint(Blueprint):
         return f'<BundleBlueprint "{self.name}">'
 
 
-class RegisterBundleTemplateFoldersHook(AppFactoryHook):
+class RegisterBundleBlueprintsHook(AppFactoryHook):
     bundle_module_name = None
     name = 'bundle_template_folders'
     run_before = ['blueprints']
@@ -88,8 +85,6 @@ class RegisterBundleTemplateFoldersHook(AppFactoryHook):
     action_table_converter = lambda bp: [bp.name, bp.template_folder]
 
     def run_hook(self, app: Flask, bundles: List[Bundle]):
-        self.babel_bundle = get_babel_bundle(bundles)
-
         for bundle_ in reversed(bundles):
             for bundle in bundle_.iter_class_hierarchy(reverse=False):
                 if (bundle.template_folder
@@ -105,6 +100,3 @@ class RegisterBundleTemplateFoldersHook(AppFactoryHook):
                                         **route.rule_options)
                     app.register_blueprint(bp)
                     self.log_action(bp)
-
-                    if self.babel_bundle:
-                        self.babel_bundle.register_blueprint(app, bp)